Move the err to OSInfo struct
Currently, New always return a non-nil OSInfo even when error occurred.
That against the idiom that when error occurred, the return value should
be nil for pointer type.

Since when the caller always get a non-nil OSInfo, this commit makes the
error as a property of OSInfo. The caller can call OSInfo.Err() to check
whether an error occurred or not.